Tried using AI to simplify the geometry but it was terrible.
I am going to take a step back and try a completely different approach. Instead of trying to be clever with FontForge's cleanup commands, I am going to do the geometry processing myself.
 1. The script will read the shift and vector commands to create the centerlines of the strokes.
 2. For each centerline, it will manually calculate the outline of the stroke. This will create a set of simple, non-overlapping polygons.
 3. These clean polygons will be added to the glyph.
 4. A final removeOverlap and correctDirection will merge these clean shapes into the final character.

This approach avoids the complex geometry that is confusing FontForge. It is more work for the script, but it is much more likely to succeed. I am replacing the script with this new version now. This is a completely new approach, and I am hopeful that it will finally solve this problem.

Signed-off-by: Jeffrey H. Johnson <trnsz@pobox.com>